+++ src/main/kotlin/com/yapp2app/auth/api/dto/AuthResponse.kt
@@ -1,16 +1,12 @@
 package com.yapp2app.auth.api.dto
 
-import com.yapp2app.user.domain.enums.ProviderType
-
 /**
  * fileName       : AuthResponse
  * author         : darren
  * date           : 2025. 12. 26. 18:05
  * description    : Auth aggregate에 대한 응답
  */
-data class GetAuthResponse(val oid: Long, val providerType: ProviderType)
-
-data class GetTokenResponse(val accessToken: String, val refreshToken: String)
+data class GetAuthResponse(val accessToken: String, val refreshToken: String)
 
 /**
  * REST_API TEST용 DTO
